global pillar for all roles
everything in the 'core' pillar otherwise I have to keep track of which
module includes which module etc etc to see which role should have which
pillar. for my homelab use-case it's not worth the time to separate that
out. normally you'd want to make sure the minions have as little pillar
data as needed.