create openstack k8s cluster with rancher api
Took some while to figure this out but apparently in the Rancher web
interface when you create a cluster it doesn't do one API call but
actually two (one to create the cluster, one to assign a nodepool to the
cluster). This is not immediately apparent because the web interface
hides this under the same one 'add cluster' page and unlike on other
pages there is no 'show API request' button in the menu here. Anyway
what turns out which is actually very nice is that you can simply press
the submit button and look at the network requests in browser dev tools and
see the payload and endpoint. This makes it so that you can click
together a pretty complex action and then simply inspect the payload
json and then use that to construct an API call with curl or whatever to
automate that step. My use-case is that I want to spawn a VM, run a
fresh rancher docker on it (no state) and then immediately have it spawn
a fresh kubernetes cluster on an (also fresh) openstack instance. In
that workflow even clicking once in a web interface is undesirable
(however nice the web-interface might be). There aren't a lot of code
examples available for this type of interaction (either in documentation
or in public projects using rancher). Deconstructing the terraform
driver gave some clues. Anyway this probably means that this isn't
exactly the intended use-case of Rancher for most people so this
approach will probably result in issues down the line.